requirements.txt
Django==5.1.1
daphne==4.1.2
psycopg[binary]==3.2.1
environs[django]==11
django-ninja==1.3.0
servestatic==1.0.0
stripe==10.12.0
django-allauth[socialaccount]==64.2.1
pgvector==0.3.3 
uvicorn==0.30.6
pytest-django==4.8.0 
pytest-asyncio==0.24.0
httpx==0.27.2
mypy==1.11
django-stubs[compatible-mypy]==5.0.4